Fix errors on switching dimensions
