+++ azalea-core/src/data_registry.rs
@@ -1,4 +1,4 @@
-use std::str::FromStr;
+use std::{io::Cursor, str::FromStr};
 
 use azalea_registry::DataRegistry;
 use simdnbt::owned::NbtCompound;
@@ -22,6 +22,26 @@ pub trait ResolvableDataRegistry: DataRegistry {
         let registry_values = registries.map.get(&name_resourcelocation)?;
         let resolved = registry_values.get_index(self.protocol_id() as usize)?;
         Some(resolved)
+    }
+
+    fn resolve_and_deserialize<'a, T: simdnbt::Deserialize>(
+        &self,
+        registries: &'a RegistryHolder,
+    ) -> Option<Result<(ResourceLocation, T), simdnbt::DeserializeError>> {
+        let (name, value) = self.resolve(registries)?;
+
+        let mut nbt_bytes = Vec::new();
+        value.write(&mut nbt_bytes);
+        let nbt_borrow_compound =
+            simdnbt::borrow::read_compound(&mut Cursor::new(&nbt_bytes)).ok()?;
+        let value = match T::from_compound((&nbt_borrow_compound).into()) {
+            Ok(value) => value,
+            Err(err) => {
+                return Some(Err(err));
+            }
+        };
+
+        Some(Ok((name.clone(), value)))
     }
 }
 impl<T: DataRegistry> ResolvableDataRegistry for T {}
