+++ azalea-world/src/container.rs
@@ -51,6 +51,7 @@ impl InstanceContainer {
         name: ResourceLocation,
         height: u32,
         min_y: i32,
+        default_registries: &RegistryHolder,
     ) -> Arc<RwLock<Instance>> {
         if let Some(existing_lock) = self.instances.get(&name).and_then(|world| world.upgrade()) {
             let existing = existing_lock.read();
@@ -72,7 +73,7 @@ impl InstanceContainer {
                 chunks: ChunkStorage::new(height, min_y),
                 entities_by_chunk: HashMap::new(),
                 entity_by_id: IntMap::default(),
-                registries: RegistryHolder::default(),
+                registries: default_registries.clone(),
             }));
             debug!("Added new instance {name}");
             self.instances.insert(name, Arc::downgrade(&world));
